+++ CHANGELOG.rst
@@ -41,6 +41,7 @@ This document describes changes between each past release.
 - Fix bug where the resource events of a request targetting two groups/collection
   from different buckets would be grouped together.
 - Fix crash when an invalid UTF-8 character is provided in URL
+- Fix crash when provided ``last_modified`` field is not divisible (e.g. string)
 
 **Internal changes**
 
@@ -59,6 +60,7 @@ This document describes changes between each past release.
   parameter anymore
 - Improved parts of the FAQ (#744)
 - Improve 404 and 403 error handling to make them customizable. (#748)
+- ``kinto.core`` resources are not schemaless by default (fixes #719)
 
 
 3.3.1 (2016-07-19)